Fix compile warnings (#6)
* fix all warnings

* cmake文件整理

+++ include/latte/net.h
@@ -93,13 +93,13 @@ class Net : public Noncopyable {
 
   inline const vector<int> &top_ids(int i) const {
     CHECK_GE(i, 0) << "Invalid layer id";
-    CHECK_LT(i, top_id_vecs_.size()) << "Invalid layer id";
+    CHECK_LT(i, static_cast<int>(top_id_vecs_.size())) << "Invalid layer id";
     return top_id_vecs_[i];
   }
 
   inline const vector<int> &bottom_ids(int i) const {
     CHECK_GE(i, 0) << "Invalid layer id";
-    CHECK_LT(i, bottom_id_vecs_.size()) << "Invalid layer id";
+    CHECK_LT(i, static_cast<int>(bottom_id_vecs_.size())) << "Invalid layer id";
     return bottom_id_vecs_[i];
   }
